refactor: TDET-59 Separate DAG definition and tasks into different files
